deploy/requirements.txt
Flask==2.0.1
google-cloud-aiplatform==1.67.1
Werkzeug==2.0.0
